+++ config.env
@@ -1,0 +1,3 @@
+spring.datasource.url=jdbc:mysql://localhost:3306/election
+spring.datasource.username=root
+spring.datasource.password=kenny
