+++ src/main/java/com/semicolon/electionsnacks/configurations/Configurations.java
@@ -1,0 +1,20 @@
+package com.semicolon.electionsnacks.configurations;
+
+
+import org.modelmapper.ModelMapper;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+
+@Configuration
+public class Configurations {
+    @Bean
+    public ModelMapper modelMapper(){
+        return new ModelMapper();
+    }
+    @Bean
+    public PasswordEncoder passwordEncoder(){
+        return new BCryptPasswordEncoder();
+    }
+}
