+++ src/main/java/com/semicolon/electionsnacks/dtos/response/ApiResponse.java
@@ -1,0 +1,21 @@
+package com.semicolon.electionsnacks.dtos.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.http.HttpStatus;
+
+import java.time.ZonedDateTime;
+
+@Getter
+@Setter
+@Builder
+@AllArgsConstructor
+public class ApiResponse {
+    private ZonedDateTime timeStamp;
+    private HttpStatus statusCode;
+    private String path;
+    private Object data;
+    private Boolean isSuccessful;
+}
