+++ src/main/java/com/semicolon/electionsnacks/exceptions/GenericException.java
@@ -1,0 +1,7 @@
+package com.semicolon.electionsnacks.exceptions;
+
+public class GenericException extends RuntimeException{
+    public GenericException(String message){
+        super(message);
+    }
+}
