+++ src/main/java/com/semicolon/electionsnacks/exceptions/GlobalExceptionHandler.java
@@ -1,0 +1,25 @@
+package com.semicolon.electionsnacks.exceptions;
+
+import com.semicolon.electionsnacks.dtos.response.ApiErrorResponse;
+import jakarta.servlet.http.HttpServletRequest;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+import java.time.ZonedDateTime;
+
+@RestControllerAdvice
+public class GlobalExceptionHandler {
+    @ExceptionHandler(GenericException.class)
+    public ResponseEntity<?> handleRegistrationException(GenericException exception, HttpServletRequest request){
+        ApiErrorResponse errorResponse = ApiErrorResponse.builder()
+                .isSuccessful(false)
+                .data(exception.getMessage())
+                .statusCode(HttpStatus.BAD_REQUEST)
+                .path(request.getRequestURI())
+                .timeStamp(ZonedDateTime.now())
+                .build();
+        return new ResponseEntity<>(errorResponse, HttpStatus.NOT_FOUND);
+    }
+}
