+++ src/main/java/com/semicolon/electionsnacks/models/ElectionType.java
@@ -1,0 +1,6 @@
+package com.semicolon.electionsnacks.models;
+
+public enum ElectionType {
+     PRESIDENTIAL,
+    GOVERNORSHIP
+}
