+++ src/main/java/com/semicolon/electionsnacks/models/Party.java
@@ -1,0 +1,21 @@
+package com.semicolon.electionsnacks.models;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import lombok.Setter;
+
+@RequiredArgsConstructor
+@Getter
+@Setter
+@Entity
+public class Party {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private String logo;
+    private String name;
+}
