+++ src/main/java/com/semicolon/electionsnacks/models/Vote.java
@@ -1,0 +1,20 @@
+package com.semicolon.electionsnacks.models;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import lombok.Setter;
+
+@Entity
+@AllArgsConstructor
+@RequiredArgsConstructor
+@Getter
+@Setter
+public class Vote {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    @ManyToOne
+    private Candidate candidate;
+}
