+++ src/main/java/com/semicolon/electionsnacks/services/EmailService.java
@@ -1,0 +1,5 @@
+package com.semicolon.electionsnacks.services;
+
+public interface EmailService {
+    void send(String to, String email);
+}
