+++ src/main/java/com/semicolon/electionsnacks/utils/PasswordEncoder.java
@@ -1,0 +1,9 @@
+package com.semicolon.electionsnacks.utils;
+
+public interface PasswordEncoder {
+    String encode(CharSequence rawPassword);
+    boolean matches(CharSequence rawPassword, String encodedPassword);
+    default boolean upgradeEncoding(String encodedPassword){
+        return false;
+    }
+}
