--- src/main/java/controller/VoterController.java
@@ -1,4 +1,0 @@
-package controller;
-
-public class VoterController {
-}
