--- src/main/java/dtos/request/EmailRequest.java
@@ -1,14 +1,0 @@
-package dtos.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@RequiredArgsConstructor
-public class EmailRequest {
-    @NotBlank(message = "This field cannot be blank") private String email;
-}
