--- src/main/java/dtos/request/UpdatePasswordRequest.java
@@ -1,16 +1,0 @@
-package dtos.request;
-
-import jakarta.validation.constraints.NotBlank;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class UpdatePasswordRequest {
-    @NotBlank(message = "This field must not be empty")
-    private String email;
-    @NotBlank(message = "This field must not be empty")
-    private String oldPassword;
-    @NotBlank(message = "This field must not be empty")
-    private String newPassword;
-}
