--- src/main/java/exceptions/GenericException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class GenericException extends RuntimeException{
-    public GenericException(String message){
-        super(message);
-    }
-}
