--- src/main/java/exceptions/GlobalExceptionHandler.java
@@ -1,4 +1,0 @@
-package exceptions;
-
-public class GlobalExceptionHandler {
-}
