--- src/main/java/models/Election.java
@@ -1,24 +1,0 @@
-package models;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import lombok.Setter;
-
-import java.time.LocalDateTime;
-
-@RequiredArgsConstructor
-@Getter
-@Setter
-@Entity
-public class Election {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
-    private ElectionType electionType;
-    private LocalDateTime startTime;
-    private LocalDateTime endTime;
-}
