--- src/main/java/models/ElectionType.java
@@ -1,6 +1,0 @@
-package models;
-
-public enum ElectionType {
-     PRESIDENTIAL,
-    GOVERNORSHIP
-}
