--- src/main/java/models/Vote.java
@@ -1,20 +1,0 @@
-package models;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import lombok.Setter;
-
-@Entity
-@AllArgsConstructor
-@RequiredArgsConstructor
-@Getter
-@Setter
-public class Vote {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
-    @ManyToOne
-    private Candidate candidate;
-}
