--- src/main/java/services/EmailService.java
@@ -1,5 +1,0 @@
-package services;
-
-public interface EmailService {
-    void send(String to, String email);
-}
