--- src/main/java/services/voters/VoterService.java
@@ -1,9 +1,0 @@
-package services.voters;
-
-import dtos.request.EmailRequest;
-import dtos.request.UpdatePasswordRequest;
-
-public interface VoterService {
-    String forgotPassword(EmailRequest emailRequest);
-    String changePassword(UpdatePasswordRequest updatePasswordRequest);
-}
