--- src/main/java/utils/PasswordEncoder.java
@@ -1,9 +1,0 @@
-package utils;
-
-public interface PasswordEncoder {
-    String encode(CharSequence rawPassword);
-    boolean matches(CharSequence rawPassword, String encodedPassword);
-    default boolean upgradeEncoding(String encodedPassword){
-        return false;
-    }
-}
