+++ src/main/resources/application.properties
@@ -1,1 +1,19 @@
+spring.datasource.url=${DATABASE_URL}
+spring.datasource.username=${DATABASE_USERNAME}
+spring.datasource.password=${DATABASE_PASSWORD}
+spring.jpa.hibernate.ddl-auto=update
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 
+server.port=9000
+
+
+spring.mail.host=localhost
+spring.mail.port=1025
+spring.mail.username=hello
+spring.mail.password=hello
+spring.mail.properties.mail.smtp.auth=true
+spring.mail.properties.mail.smtp.ssl.trust="*"
+spring.mail.properties.mail.smtp.starttls.enable=true
+spring.mail.properties.mail.smtp.connectiontimeout=5000
+spring.mail.properties.mail.smtp.timeout=3000
+spring.mail.properties.mail.smtp.writetimeout=5000
