+++ src/test/java/ElectionSnacksApplicationTests.java
@@ -1,0 +1,11 @@
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ElectionSnacksApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
