--- src/test/java/com/semicolon/electionsnacks/ElectionSnacksApplicationTests.java
@@ -1,13 +1,0 @@
-package com.semicolon.electionsnacks;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ElectionSnacksApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
